changes requested from PR
