+++ README.md
@@ -2,7 +2,7 @@
 
 Code Coverage
 
-![Statements](https://img.shields.io/badge/statements-79.65%25-red.svg?style=flat) ![Branches](https://img.shields.io/badge/branches-87.01%25-yellow.svg?style=flat) ![Functions](https://img.shields.io/badge/functions-81.5%25-yellow.svg?style=flat) ![Lines](https://img.shields.io/badge/lines-79.65%25-red.svg?style=flat)
+![Statements](https://img.shields.io/badge/statements-79.59%25-red.svg?style=flat) ![Branches](https://img.shields.io/badge/branches-87.05%25-yellow.svg?style=flat) ![Functions](https://img.shields.io/badge/functions-81.5%25-yellow.svg?style=flat) ![Lines](https://img.shields.io/badge/lines-79.59%25-red.svg?style=flat)
 
 
 ## Introduction
