+++ src/dwn.ts
@@ -2,7 +2,7 @@ import type { BaseMessage, RequestSchema } from './core/types';
 import type { DidMethodResolver } from './did/did-resolver';
 import type { Interface, MethodHandler } from './interfaces/types';
 import type { MessageStore } from './store/message-store';
-import * as decoder from '../src/utils/decoder';
+import * as encoder from '../src/utils/encoder';
 import { addSchema } from './validation/validator';
 import { CollectionsInterface, PermissionsInterface } from './interfaces';
 import { DidKeyResolver } from './did/did-key-resolver';
@@ -63,7 +63,7 @@ export class Dwn {
   async processRequest(rawRequest: Uint8Array): Promise<Response> {
     let request: RequestSchema;
     try {
-      const requestString = decoder.bytesToString(rawRequest);
+      const requestString = encoder.bytesToString(rawRequest);
       request = JSON.parse(requestString);
     } catch {
       throw new Error('expected request to be valid JSON');
