--- src/utils/decoder.ts
@@ -1,14 +1,0 @@
-import { base64url } from 'multiformats/bases/base64';
-
-
-
-const textDecoder = new TextDecoder();
-
-export function bytesToString(content: Uint8Array): string {
-  const bytes = textDecoder.decode(content);
-  return bytes;
-}
-export function base64urlToBytes(base64urlString: string): Uint8Array {
-  const content = base64url.baseDecode(base64urlString);
-  return content;
-}
